Rename docs/built-on-ethereum/open-finance/what-is-open-finance.md to docs/built-on-ethereum/open-finance/web3js/what-is-open-finance.md
title: Decentralized Finance on Ethereum - EthHub

description: Decentralized Finance refers to a number of decentralized protocols building open financial infrastructure.

# Decentralized Finance

## Summary

Decentralized Finance (a.k.a. "DeFi" or "Open Finance") refers to a number of decentralized protocols building open financial infrastructure. These protocols are valuable because they're creating the necessary plumbing to enable anyone in the world with an internet connection to access self-sovereign, censorship resistant financial services.

In the existing system, all financial services are controlled by a central party. Whether it's basic money transfers, asset purchases, or lending, you must go through an intermediary who charges rent for mediating financial transactions. Ethereum-based financial services, on the other hand, connect individuals peer-to-peer and allow them to access basic financing more easily and affordably.

## How is all this possible?

A big reason the financial services industry is as bloated as it is today is because financial transactions in the digital realm are not secure. We have no method of trusting our online counterparties so we pay fees to financial institutions who enforce that trust.

Ethereum leverages the same principles that create 'digital trust' within Bitcoin, and applies to them to smart contracts, self-executing pieces of code that execute business logic after some predefined conditions are met. Smart contracts look a lot like financial contracts in that they escrow funds and transfer them around in response to certain events. This system is superior because the encoded business logic can't be manipulated by a central party once its deployed to the Ethereum mainnet.

## Blockchain based financial services are superior to their centralized counterparts across the following primary features:

1. **Permissionless**: An internet connection is all you need to access these services
2. **Censorship Resistance**: No central party is able to reverse the order of transactions and turn off the service
3. **Trustless**: Users don't have to trust a central party to ensure that transactions are valid
4. **Transparent**: Public blockchains like Ethereum are completely transparent and auditable
5. **Programmable**: Developers can create and intertwine financial services at a very low cost
6. **Efficiency**:  Open financial services are powered by code, not humans, and as such entail much lower middleman costs, if at all

## Decentralized Finance is generally broken up into the following categories:

1. Asset Exchange
2. Credit Scoring
3. Insurance
4. Indexing/Baskets
5. Lending
6. Derivatives
7. Stablecoins
8. Smart Securities
9. Prediction Markets
10. Open Source Banking Applications

## Resources

* [Decentralized Finance dApps on Ethereum](https://defiprime.com/ethereum)